Better message validation.
Adding "Alumnus" changed the message validation because we had to support "a Senior"
or "an Alumnus". I fixed it, but not in a great way—you had to put "a(n)" in the message, which
was sort of lame.

I think it's better now, but the sentence still reads weird for alums. Maybe we switch major and class standing?